splitting sops_read_data into multiple files, added validateInputType and tests for validateInputType, checking for valid input in data_sops_external
